# openapi.yaml
openapi: 3.0.3
info:
  title: CrewSwap Mobile App API
  version: 1.0.0
  description: |
    OpenAPI specification for CrewSwap mobile and admin API.
    All endpoints return a consistent JSON envelope.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://127.0.0.1:8000/api/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /reports:
    get:
      tags: [Reports]
      summary: List reports
      parameters:
        - name: reporter_id
          in: query
          schema: { type: integer }
        - name: defendant_id
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string, enum: [waiting, solved, rejected, pending, reviewed, resolved] }
        - name: type
          in: query
          schema: { type: string, enum: [spam, bad_language, disrespect, wrong_information] }
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/per_page'
      responses:
        '200':
          description: Reports list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Report'
    post:
      tags: [Reports]
      summary: Create report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          success: { type: boolean, example: true }
                          report_id: { type: integer, example: 401 }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    per_page:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed
        data:
          type: object
          additionalProperties: true
        meta:
          $ref: '#/components/schemas/Meta'
    PaginatedEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: List fetched
        data:
          type: object
          additionalProperties: true
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
            - type: object
              properties:
                pagination:
                  $ref: '#/components/schemas/Pagination'
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        per_page: { type: integer, example: 20 }
        total_count: { type: integer, example: 245 }
        total_pages: { type: integer, example: 13 }
...
    Report:
      type: object
      properties:
        id: { type: integer }
        reporter_id: { type: integer }
        defendant_id: { type: integer }
        type: { type: string }
        message: { type: string }
        status: { type: string }
        report_date: { type: string, format: date-time }
    CreateReportRequest:
      type: object
      required: [reporter_id, defendant_id, type, message]
      properties:
        reporter_id: { type: integer }
        defendant_id: { type: integer }
        type: { type: string, enum: [spam, bad_language, disrespect, wrong_information] }
        message: { type: string }
...